requirements.txt
flask
openai==0.28
beautifulsoup4
requests
python-dotenv